pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bio-debug-crispr"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "anthropic",
    "python-dotenv",
    "streamlit"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "tools*", "prompts*", "utils*", "main*", "app*"]

[project.scripts]
run_app = "main:main"
engineer = "agents.engineer:main"
